+++ OneBlink.SDK/models/AssetUploadCredentialsRequest.cs
@@ -1,8 +1,0 @@
-namespace OneBlink.SDK.Model
-{
-    public class AssetUploadCredentialsRequest
-    {
-        public string assetPath { get; set; }
-        public string organisationId { get; set; }
-    }
-}