+++ OneBlink.SDK/models/AttachmentData.cs
@@ -1,13 +1,30 @@
 namespace OneBlink.SDK.Model
 {
-    public class AttachmentData
+    public class AttachmentData : S3UploadResponse
     {
-        public string id {get;set;}
-        public string contentType {get;set;}
-        public string fileName {get;set;}
-        public bool isPrivate {get;set;}
-        public string url {get;set;}
-        public S3Details s3 { get; set; }
-        public string uploadedAt {get; set;}
+        public string id
+        {
+            get; set;
+        }
+        public string contentType
+        {
+            get; set;
+        }
+        public string fileName
+        {
+            get; set;
+        }
+        public bool isPrivate
+        {
+            get; set;
+        }
+        public string url
+        {
+            get; set;
+        }
+        public string uploadedAt
+        {
+            get; set;
+        }
     }
 }