+++ OneBlink.SDK/models/EmailAttachmentData.cs
@@ -1,11 +1,7 @@
 namespace OneBlink.SDK.Model
 {
-    public class EmailAttachmentData
+    public class EmailAttachmentData : S3UploadResponse
     {
-        public S3Details s3
-        {
-            get; set;
-        }
         public string contentType
         {
             get; set;