+++ OneBlink.SDK/models/S3Details.cs
@@ -1,9 +1,0 @@
-namespace OneBlink.SDK.Model
-{
-    public class S3Details
-    {
-        public string region { get; set; }
-        public string bucket { get; set; }
-        public string key { get; set; }
-    }
-}