+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- uploads to go storage service
+
 ## [5.6.1] - 2024-04-10
 
 ### Added